+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.3 (unreleased)
-Nothing so far
+
+#### Additons
+* LF/AD-5.5
 
 ## 0.2.2